requirements.txt
cassandra-driver==3.25.0
fastapi==0.68.0
pytest-6.2.5
uvicorn==0.14.0
